requirements.txt
python-binance>=1.0.19
Flask>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
requests>=2.31.0
plotly>=5.17.0  # For interactive charts
waitress>=3.0.0  # For production WSGI server